feat(quality): shared anchor generators (extract gen_labeled + crm_df)
T1: one definition of each anchor shape in scripts/autoconfig_quality/anchors.py
(gen_labeled, crm_df, re-exported make_healthcare_df). Redirected the two donor
tests (test_quality_gate, test_autoconfig_multisource); test_autoconfig_benchmarks
has NO gen_labeled copy (reviewer's claim was inaccurate). 30 tests pass.
