+++ packages/python/goldenmatch/tests/test_quality_gate.py
@@ -24,13 +24,17 @@ ground-truth pairs by construction. Deterministic seed.
 """
 from __future__ import annotations
 
-import random
-from collections import defaultdict
-from itertools import combinations
+import sys
+from pathlib import Path
 
-import polars as pl
 import pytest
 from goldenmatch import dedupe_df
+
+# gen_labeled is the shared person-match anchor (one definition for tests + the
+# quality harness). Add repo-root to sys.path so scripts.* is importable.
+_REPO_ROOT = Path(__file__).resolve().parents[4]
+if str(_REPO_ROOT) not in sys.path:
+    sys.path.insert(0, str(_REPO_ROOT))
 from goldenmatch.config.schemas import (
     BlockingConfig,
     BlockingKeyConfig,
@@ -40,53 +44,7 @@ from goldenmatch.config.schemas import (
 )
 from goldenmatch.core.evaluate import evaluate_clusters
 
-_SURN = [
-    "Smith", "Jones", "Williams", "Brown", "Davis", "Miller", "Wilson", "Moore",
-    "Taylor", "Anderson", "Thomas", "Jackson", "White", "Harris", "Martin",
-    "Thompson", "Garcia", "Martinez", "Robinson", "Clark", "Rodriguez", "Lewis",
-    "Lee", "Walker", "Hall", "Allen", "Young", "King", "Wright", "Lopez",
-]
-_FIRST = [
-    "Alex", "Blair", "Casey", "Dana", "Eli", "Finley", "Gray", "Harper",
-    "Indigo", "Jamie", "Kendall", "Logan", "Morgan", "Noel", "Oakley", "Parker",
-    "Quinn", "Riley", "Sage", "Taylor", "Umi", "Val", "Wren", "Xena", "Yael",
-    "Zane", "Avery", "Brook", "Cleo", "Drew",
-]
-
-
-def _typo(s: str, rng: random.Random) -> str:
-    if len(s) < 3:
-        return s
-    i = rng.randrange(len(s) - 1)
-    return s[:i] + s[i + 1] + s[i] + s[i + 2:]  # adjacent-char swap
-
-
-def gen_labeled(n_entities: int = 400, seed: int = 7) -> tuple[pl.DataFrame, set]:
-    """Synthetic records with known ground truth. Each entity = 1 original +
-    0-2 typo'd clones (sharing email + zip). Returns (df, ground_truth_pairs)
-    where pairs are (row_index, row_index) for rows of the same true entity."""
-    rng = random.Random(seed)
-    n_zip = max(1, n_entities // 2)
-    tagged: list[tuple[dict, int]] = []
-    for e in range(n_entities):
-        f, l = rng.choice(_FIRST), rng.choice(_SURN)
-        z = f"{rng.randrange(n_zip):05d}"
-        email = f"{f}.{l}.{e}@x.com".lower()
-        tagged.append(({"first_name": f, "last_name": l, "email": email, "zip": z}, e))
-        for _ in range(rng.choice([0, 0, 1, 1, 2])):
-            tagged.append(
-                ({"first_name": _typo(f, rng), "last_name": l, "email": email, "zip": z}, e)
-            )
-    rng.shuffle(tagged)
-    df = pl.DataFrame([rec for rec, _ in tagged])
-    by_entity: dict[int, list[int]] = defaultdict(list)
-    for pos, (_, e) in enumerate(tagged):
-        by_entity[e].append(pos)
-    gt: set = set()
-    for positions in by_entity.values():
-        for a, b in combinations(sorted(positions), 2):
-            gt.add((a, b))
-    return df, gt
+from scripts.autoconfig_quality.anchors import gen_labeled  # noqa: E402
 
 
 def _cfg(backend: str | None, kind: str) -> GoldenMatchConfig:
