+++ scripts/autoconfig_quality/__init__.py
@@ -1,0 +1,6 @@
+"""Auto-config quality harness — run auto-config across a corpus (real labeled
+datasets + synthetic failure-shape anchors) and emit a comparative scorecard
+(fast config-signals + F1) with a committed baseline, diff, gate, and bless.
+
+See docs/superpowers/specs/2026-06-22-autoconfig-quality-harness-design.md.
+"""
